ICMSLST-1174 Rename the permission `reference_data_access` to `ilb_admin`
This permission is used to give access to members of the ILB team (Import
Licensing Branch). "Reference data access" referred to editing the various
lookup fields (choices), but as it turns out the app does not distinguish
between being able to configure these data vs managing the rest of the app.