ICMSLST-719 Prevent double submitting <form>s on case import / export pages
Adds a new js file which looks for all elements with `class="no-double-submit"`
on page load. For matching elements with the attribute `method="POST"`,
install a "submit" event handler (i.e. form submit, either by clicking a button
or keyboard entry, etc.).

The submit handler checks the form for a special class ("double-submit-blocked")
and if the form already has the class, then the submit event is cancelled. Else
the special class is added, and the submit event happens as normal.