ICMSLST-1029 Link to create a certificate application from a template
Added a link to create an application in the list of the user's saved
application templates.

The "/export/?from-template=[ID]" path will redirect to the specific type
depending on the saved template. If the template ID is invalid, or if the
user is not the owner of the template, then the user sees the normal page
for choosing an application type.

(Restricting template use to the owner prevents leaking PII - when we
implement sharing a template with other users this policy will need to
be updated.)

The certificate application edit form views now accept a "?from-template"
query parameter and on GET (but not POST) will use the template to set
initial values for the form data.

Dummy templates have been added for the ilb_admin and exporter_user
users, for the different application types.