ICMSLST-1190 Allow editing of basic application template details
Added a new route "/cat/edit/[pk]/[step]/" for editing the template data for
use with a certificate application template model. The view requires the
same permissions as the template creation view, and checks that the user has
edit permissions for that template (currently requires the user is the owner
of the template).

Choice fields (select inputs) return their data as Django models, which are
serialized as the object's primary key. This is different to Django's own
serializer framework, which represents a model as a dict/object with the key,
the app name and refs for foreign keys.