ICMSLST-1178 Archiving for certificate application templates
Also ICMSLST-1204, restoring a previously archived template.

This changes the /cat/ template list view so that for each template in
the list there is an action to archive or restore that template, which
updates the CertificateApplicationTemplate.is_active boolean field
appropriately.

The queryset manager for .objects.active() and .objects.inactive() will
be useful when we implement the filtering feature for the template list.