ICMSLST-1221 Prevent creating an application from an archived template
- In the list of templates ("/cat/"), for an archived template, don't show the
"Create application" or "Edit template" actions.
- In the list of templates ("/cat/") show only current templates by default.
- In the application from template view ("/export/create/[type]/template/[id]"),
check if the template is archived, and if it is then redirect the user to the
start application view and show a message saying they cannot use an archived
template.

I removed the CertificateApplicationTemplate.objects.active() / .inactive()
manager because it is not used. I originally added it because I assumed we
would want it to filter the list of templates on the "create export app"
view, but as implemented it serves no purpose.