ICMSLST-1269 Stub methods for CHIEF integration (#699)
The path for the call-back is "/chief/license-data-callback".

New settings for configuring Hawk authentication:

- CHIEF_LICENSE_URL, the icms-hmrc API to send license requests.
- HAWK_AUTH_ID, client ID shared by icms and icms-hmrc (not a secret).
- HAWK_AUTH_KEY, signing key shared by icms and icms-hmrc (secret).

Both HAWK_AUTH_ID and HAWK_AUTH_KEY must be shared with the sender
(in our case, the sender is icms-hmrc).

When sending data to the call-back, we must return either a 207 multi
status or 208 already reported status. For a 207 response, the body
is a JSON object with keys for "accepted" and "rejected", and with
values a list of objects like `{"id": <license-id>}`.