ICMSLST-982 Add all countries to deactivated firearms form choices
The "Firearms and Ammunition (Deactivated) Issuing Countries" country group
is used on the "Firearms and Ammunition (Deactivated Firearms Licence)" form
and its related "add goods certificate" form.

On the legacy ICMSv1 form we have "Burma (Myanmar)", but I changed that to
"Myanmar (Burma)" across the app because that's the actual country name
these days.