ICMSLST-1388 New command "./manage.py countries export"
The command writes the country records and country group names to <stdout>
in CSV format. Where a country belongs to a group, there is a "Y" in the
cell (the group names are the column headers).

Use this like "./manage.py countries export > report.csv" to save the
output to a file named "report.csv".

There is also a "./manage.py countries import" command that reads from
<stdin>, but all it does at the moment is print how many countries and
groups were in the CSV.